CONFIG_PAGE_SIZE to 16 KB, disable kernel debug.
CONFIG_PAGE_SIZE on MIPS can be 16 KB or 64 KB. But some programs are
not deal it correctly, e.g. hardcoded page size to 16 KB. Although
64 KB can boost the performance, we have to switch back to 16 KB
to avoid breaking programs.

CONFIG_CRASH_DUMP is useless, in faect, without other necessary config,
even enable it, we can't debug the kernel. So disable it.